make a sign up and login set cookies, and separate the processes (login, signup, setCookies) each of them in own file #68
